Add multi-console log support
TODO: Add tests

To use this instead of having a console logfile of the name
$CONSOLE_LOGS/$SYSTEM_NAME instead have a directory of the name
$CONSOLE_LOGS/$SYSTEM_NAME/ and inside of that directory have console
logfiles.

For example if CONSOLE_LOGS=/var/consoles/ (which is the default) and
you have a system with the name 'beaker-test-vm1.beaker' then create a
directory:
  mkdir /var/consoles/beaker-test-vm1.beaker/

And put your console log files in that directory.